+++ R/example/example_dependence_plot.R
@@ -1,6 +1,44 @@
 # **SHAP dependence plot**
-shap.plot.dependence.color(data_long = shap_long_iris, x="Petal.Length",
-                           y = "Petal.Width", color_feature = "Petal.Width")
-# the without color version, just plot SHAP value against feature value:
-shap.plot.dependence(data_long = shap_long_iris, "Petal.Length")
 
+# 1. simple dependence plot with SHAP values of x on the y axis
+shap.plot.dependence(data_long = shap_long_iris, x="Petal.Length", add_hist = TRUE)
+
+# 2. can choose a different SHAP values on the y axis
+shap.plot.dependence(data_long = shap_long_iris, x="Petal.Length",
+                           y = "Petal.Width")
+
+# 3. color by another feature's feature values
+shap.plot.dependence(data_long = shap_long_iris, x="Petal.Length",
+                           color_feature = "Petal.Width")
+
+# 4. choose 3 different variables for x, y, and color
+shap.plot.dependence(data_long = shap_long_iris, x="Petal.Length",
+                           y = "Petal.Width", color_feature = "Petal.Width")
+
+# Optional to add hist or remove smooth line, optional to plot fewer data (make plot quicker)
+shap.plot.dependence(data_long = shap_long_iris, x="Petal.Length",
+                     y = "Petal.Width", color_feature = "Petal.Width",
+                     add_hist = TRUE, smooth = FALSE, dilute = 3)
+
+# to make a list of plot
+plot_list <- lapply(names(iris)[2:3], shap.plot.dependence, data_long = shap_long_iris)
+
+# **SHAP interaction effect plot **
+
+# To get the interaction SHAP dataset for plotting, need to get `shap_int` first:
+mod1 = xgboost::xgboost(
+  data = as.matrix(iris[,-5]), label = iris$Species,
+  gamma = 0, eta = 1, lambda = 0,nrounds = 1, verbose = FALSE)
+# Use either:
+data_int <- shap.prep.interaction(xgb_mod = mod1,
+                                  X_train = as.matrix(iris[,-5]))
+# or:
+shap_int <- predict(mod1, as.matrix(iris[,-5]),
+                    predinteraction = TRUE)
+
+# if data_int is supplied, y axis will plot the interaction values of y (vs. x)
+shap.plot.dependence(data_long = shap_long_iris,
+                           data_int = shap_int_iris,
+                           x="Petal.Length",
+                           y = "Petal.Width",
+                           color_feature = "Petal.Width")