+++ NEWS.md
@@ -2,4 +2,5 @@
 
 * Added a `NEWS.md` file to track changes to the package.
 * 07/30/2019 Version 0.0.1 uploaded to cran.
-* 08/10/2019 Version 0.0.2 released with update. 
+* 08/10/2019 Version 0.0.2 released with update, fixed some bugs with dilute. 
+* 08/27/2019 Major change, merge function `shap.plot.dependence.color` into `shap.plot.dependence`.